requirements.txt
fastapi
uvicorn
requests
python-dotenv
pydantic
motor
pydantic-settings
PyJWT